added drawing V1

+++ 1. MicroMouse Code/0. MicroMouse - STM32F407VGT6/PCB_V1/Core/Inc/AML_Keyboard.h
@@ -10,7 +10,7 @@
 
 
 
-GPIO_PinState AML_Keyboard_GetKey(short key);
+GPIO_PinState AML_Keyboard_GetKey(uint8_t key);
